RUM-16417: Move RemoteConfigurationForgeryFactory to core testFixtures
