+++ dd-sdk-android-core/src/test/kotlin/com/datadog/android/utils/forge/Configurator.kt
@@ -70,9 +70,6 @@ internal class Configurator :
 
         forge.addFactory(PersistenceStrategyBatchForgeryFactory())
 
-        // Remote Config
-        forge.addFactory(RemoteConfigurationForgeryFactory())
-
         forge.useJvmFactories()
 
         // telemetry
